refactor common code into core module
