+++ kafka-junit-core/src/main/java/com/salesforce/kafka/test/KafkaTestServer.java
@@ -23,7 +23,7 @@
  * USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.salesforce.kafka.test.junit4;
+package com.salesforce.kafka.test;
 
 import com.google.common.collect.Maps;
 import com.google.common.io.Files;
