+++ kafka-junit5/src/main/java/com/salesforce/kafka/test/junit5/SharedKafkaTestResource.java
@@ -25,6 +25,8 @@
 
 package com.salesforce.kafka.test.junit5;
 
+import com.salesforce.kafka.test.KafkaTestServer;
+import com.salesforce.kafka.test.KafkaTestUtils;
 import org.apache.curator.test.TestingServer;
 
 /**
