+++ pom.xml
@@ -30,10 +30,13 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.salesforce.kafka.com.salesforce.kafka.test.junit5.test</groupId>
+    <groupId>com.salesforce.kafka.com.salesforce.kafka.test</groupId>
     <artifactId>kafka-junit</artifactId>
-    <version>2.0.0-SNAPSHOT</version>
+    <version>2.0.0</version>
+
+    <!-- Submodules -->
     <modules>
+        <module>kafka-junit-core</module>
         <module>kafka-junit4</module>
         <module>kafka-junit5</module>
     </modules>
